+++ packages/svelte/tests/snapshot/samples/async-top-level-inspect-server/_expected/client/index.svelte.js
@@ -6,7 +6,7 @@ var root = $.from_html(`<p> </p>`);
 
 export default function Async_top_level_inspect_server($$anchor) {
 	var data;
-	var $$promises = $.run([async () => data = await Promise.resolve(42),,]);
+	var $$promises = $.run([async () => data = await Promise.resolve(42), () => void 0]);
 	var p = root();
 	var text = $.child(p, true);